+++ examples/i18n/middleware.ts
@@ -4,6 +4,7 @@ import { createI18nMiddleware } from 'fumadocs-core/middleware';
 export default createI18nMiddleware({
   languages,
   defaultLanguage,
+  hideLocale: 'default-locale',
 });
 
 export const config = {