+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-next-docs-app
 
+## 12.3.0
+
+## 12.2.5
+
 ## 12.2.4
 
 ### Patch Changes